+++ src/extension.ts
@@ -4,6 +4,7 @@ import { getColors } from './colors';
 import * as commands from './commands';
 import { provideDefinition, provideEventHover } from './definitions';
 import * as fs from 'fs';
+import { Config } from './config';
 
 export function activate(context: vscode.ExtensionContext) {
 	let ws = vscode.workspace.workspaceFolders;
@@ -14,12 +15,33 @@ export function activate(context: vscode.ExtensionContext) {
 	let iconPath = workspace.with({
 		path: workspace.fsPath + "/" + fs.readdirSync(workspace.fsPath).find((item) => item.startsWith("icon-exports-"))
 	});
-	console.log(iconPath);
+
 	let fluidAttributes = workspace.with({ path: workspace.path + "/.vscode/fluid-attributes.json" });
 	let tagsPath = workspace.with({ path: workspace.path + "/.vscode/item-tag-attributes.json" });
 	let langPath = workspace.with({ path: workspace.path + "/.vscode/lang-keys.json" });
 	let groupsPath = workspace.with({ path: workspace.path + "/local/kubejs/event_groups" });
 
+	let config = workspace.with({ path: workspace.path + "/kubejs/config/probejs.json" });
+	let configData = Config.fromData(JSON.parse(fs.readFileSync(config.fsPath, "utf-8")));
+	console.log(configData);
+	if (!configData.enabled) { return; }
+
+	if (configData.isInteractiveValid()) {
+		if (configData.interactiveMode === 0) {
+			configData.interactiveMode = 1;
+			vscode.window.showInformationMessage("ProbeJS: Interactive mode is enabled. You can now use the REPL after reloading game.");
+			fs.writeFileSync(config.fsPath, JSON.stringify(configData.overwrite(JSON.parse(fs.readFileSync(config.fsPath, "utf-8"))), null, 4));
+		}
+
+		// won't start if mode is not 1 e.g. manually disabled
+		if (configData.interactiveMode === 1) {
+			context.subscriptions.push(
+				commands.createRepl(context, configData.port)
+			);
+		}
+	}
+
+	// TODO: Read config to get the ws server status and port
 	const collector = new Collector();
 	let collected = true;
 	collector.collectIcons(iconPath);
@@ -48,7 +70,6 @@ export function activate(context: vscode.ExtensionContext) {
 				}
 
 				return collector.getHover(word, workspace);
-
 			}
 		}));
 	});
@@ -60,6 +81,13 @@ export function activate(context: vscode.ExtensionContext) {
 	if (!collected) {
 		// warn vscode user that the attributes file is not found
 		vscode.window.showWarningMessage("ProbeJS: Some attribute files are not found. Things might not work.");
+	}
+
+	if (collected) {
+		context.subscriptions.push(
+			vscode.workspace.createFileSystemWatcher(workspace.fsPath + "/.vscode/*").onDidChange(() => {
+				vscode.commands.executeCommand("probejs.reload");
+			}));
 	}
 
 	context.subscriptions.push(